Update to main.py
Add error messages.
Added some information messages.

+++ main.py
@@ -3,7 +3,7 @@ import youtube_dl
 import os
 from discord.ext import commands
 
-TOKEN = 'TOKEN HERE!'
+TOKEN = 'ODY0NDk1NTUyODM0MTc0OTg2.YO2SOQ.nnNRPX5EjQckUSr0JE_nCktwirk'
 
 client = commands.Bot(command_prefix='ydl!')
 
@@ -35,7 +35,25 @@ async def DL(ctx, url=None):
         #print(file_info)
 
         await ctx.channel.send('ダウンロード完了', file=discord.File(f'./cache/{file_info[0]}'))
+        print('[INFO]アップロード完了')
 
-    os.remove(f'./cache/{file_info[0]}')
+        os.remove(f'./cache/{file_info[0]}')
+        print('[INFO]キャッシュを削除しました。')
+
+
+@client.event
+async def on_command_error(ctx, error):
+    if isinstance(error, commands.CommandInvokeError):
+        sentence = 'エラー:容量が超過しているため。アップロードできませんでした…'
+
+        await ctx.channel.send(sentence)
+        print(f'[INFO Error]{sentence}')
+        
+        path = './cache'
+        file_info = os.listdir(path)
+        print('[INFO]キャッシュを調べています。')
+        
+        os.remove(f'./cache/{file_info[0]}')
+        print('[INFO]キャッシュを削除しました。')
 
 client.run(str(TOKEN))
